Add tests demonstrating issue with isExtensionMethod and suspend funs
Currently, PsiMethodItem.isExtensionMethod considers all suspend funs to
be extension methods because it compares the number of parameters on the
MethodItem to the number of parameters in source, and if there are more
on the item, assumes the extra parameter is the receiver. However,
suspend funs also get an extra continuation parameter compared to
source, so they are incorrectly being considered extensions.

This adds a model test to show the behavior, as well as tests for the
lint checks that currently use isExtensionMethod.

Bug: 425903432
Test: adds tests -- baselined for snapshots because snapshots always return false for isExtensionMethod

Change-Id: I5440b2107cfc830452cc3acfa4f356cf5cb92270