Track bytecode constructors with DefaultConstructorMarker when needed
The compiler generates extra overloads of kotlin source constructors with a DefaultConstructorMarker parameter added (if a constructor has an optional parameter, there is also an additional overload with both an int and DefaultConstructorMarker parameter added). In general, these aren't important for API tracking since we can just do compatibility checks with the regular version without the DefaultConstructorMarker parameter.

If a constructor uses a value class type, it can only be used from Kotlin source. The only version of the constructor that will exist in bytecode is the DefaultConstructorMarker one, so in that case it makes sense to track it for binary compatibility checks.

This adds DefaultConstructorMarker constructors to the codebase when there isn't already a constructor present with a matching signature excluding the DefaultConstructorMarker parameter, since these will be the cases where the DefaultConstructorMarker constructor is needed for binary compatibility checks.

Bug: 417740481
Test: CommonTargetLanguageTest

Change-Id: I65f9487c4b0bc82ccac8f6fa121aac180c4291a9