Update property Deprecated(HIDDEN) test to include compiled sources
Due to K1 issues, this test leaves out some accessors which are incorrectly not generated in psi. A followup change adds compiled sources for this test to show the impact of the change, but that also adds in the previously excluded accessors. Add the compiled sources so that the diff in the followup is just from the code changes.

Bug: 428197925
Test: updates test
Change-Id: Id2e80afb7c12c8c0a2a741705fe1dad5df4bee78