Remove value class accessor workarounds
Now that we aren't generating property accessors from fake UAST, we can get rid of some workarounds that were needed before to link value class type properties with their accessors.

In the case that JvmName is used on the accessors (they are accessible from Java in this case), using the inlined type to find them is still necessary.

Bug: 428990990
Test: Updated test of JvmName usage on accessors. The test is failing for the snapshot codebase because the property attached to a MethodItem isn't included by the snapshot taker.

Change-Id: I05eb589addcda522a339c8e4769bef30a4fc2ea1