Propagate hidden status from fields to properties
If a property with an exposed backing field (const or JvmField) is annotated with `@field:RestrictTo` or `@RestrictTo` (with no target), the compiler applies the annotation to the field, and the RestrictTo lint check flags references to the property from both Java and Kotlin clients.

Right now, metalava is applying the annotation to the property when no annotation target is used, but not when the `field` target is used, even though the compiler and lint check treat these the same way. This propagates hidden status from a backing field to its property to match the behavior of the RestrictTo check.

This also makes sure that fields and properties are correctly linked together when the property is in a companion and the backing field is in the containing class, which is the case for const properties.

Bug: 417262373
Test: HideAnnotationTest
Change-Id: I52c2fd6c2553c74653653bdfd06a8fd57dc8454d