Add and create simple model for Javadoc description
This change adds a simple model for javadoc content, i.e. text and
inline tags, and constructs it from the objects created by
`JavadocParser` that were previously just discarded. It updates the
`JavadocParserTest` to create a string representation of the structure
and verifies the structure as part of each test.

Bug: 429965593
Test: ./gradlew
Change-Id: I6d578c7328017338456f212ae859a4422d50d18e