Update metalava tests to include jars
These tests use deprecation level hidden or value class types, and method generation will be updated for those in the followups. This adds jars to the tests, for bytecode API generation, so it is clear what the impact of the followup changes are.

This also makes some modifications to the code samples used in tests so they would compile.

Bug: 421204393
Bug: 421201575
Test: updates tests
Change-Id: Ie887d06f3a6920583f942092a841fd98796be64e