Update metalava to AGP 8.13.0-alpha03 / lint 31.13.0-alpha03
This update deprecates KaTypeNullability, so usages has been replaced with the `isNullable`, `isMarkedNullable`, and `hasFlexibleNullability` properties. The usage of these properties is explained in code comments.

The update also fixes an issue with annotations not being applied to compiler-generated no-args constructors. This impacts a few AndroidX constructors (see aosp/3717167), requiring androidx_with_metalava to be disabled.

Also reenables the androidx-studio-integration target, which had been disabled pending this version update.

Bug: 434743977
Bug: 417687416
Test: ./gradlew test, updates a test which was previously demonstating incorrect K2 behavior which has been fixed
Change-Id: I7d3288b62f358dccfdbf4d8d170c956f709aa055