Generate delegate functions with kotlin analysis API
Originally, functions from a kotlin delegate were not tracked by metalava at all because they are not included in psi. When bytecode API tracking was added, the functions started being generated as bytecode-only. This starts generating the functions with the kotlin analysis API, which means that now the functions will be listed as InaccessibleFromJava. This still isn't right, since the methods can be accessed from Java, but it is an improvement from the previous state because now source checks will be done for the delegated functions as they are tracked as usable from Kotlin source.

Bug: 421201575
Bug: 323223405
Test: ApiFileTest, AndroidX signature file changes in aosp/3720870
Change-Id: Ifef8345ff2d4b896db9665bff0c74584cfe60543