Integrate DocCommentParser into AbstractItemDocumentation
Previously, `AbstractItemDocumentation` simply searched `text` for
`@hide`, `@doconly` and `@removed`. This change switches it to use
`DocComment` to check for the presence of block tags of type
`hide`, `doconly` and `removed` respectively.

The latter does not behave exactly the same as the former. However,
thanks to the special handling added in a preceding change this does
work correctly on Android.

Bug: 429965593
Test: ./gradlew
      # Before applying this change to Android ran:
      tools/metalava/scripts/gather-android-metalava-artifacts.py before
      # After applying this change to Android ran:
      tools/metalava/scripts/gather-android-metalava-artifacts.py after
      # Made sure that there are no differences.
      meld before after
Change-Id: Ie881dbb7052b735d1b425315efa8257fa027ff0e