Create type aliases with analysis api
Currently, type aliases are created through kt psi. This updates it to instead use the analysis API, as a more direct representation of how the type alias is defined in Kotlin source, and means type aliases can be generated when using K1 now.

Right now there are some transformations of expect/actual annotations that rely on type aliases being created before anything else in the codebase, but the rest of the analysis API creation steps happen after items are created with psi, so that required breaking up the KaCodebaseAssembler steps to keep the same ordering.

Bug: 135191699
Test: ./gradlew test, all tests with K1 are removed from the baseline files because they pass now

Change-Id: I5721308f244fd73c7c1fce42729a759d77f7e971