Update generation of methods annotated JvmName
If a method is annotated with JvmName, that usually means there should be two signatures generated for it: one with the source name that can only be used from Kotlin, and one with the jvm name which can be used from Java and bytecode. This updates the generation of the methods to reflect that, and also handles the case where even though JvmName is used the Kotlin and Java signatures end up being the same.

This doesn't update kotlin property accessors which use JvmName, because regardless of whether JvmName is used, kotlin property accessor methods can't be resolved in kotlin source (only direct property access can be used), and it would make more sense to update all accessors at once if we're going to change the tracking of those rather than treating JvmName accessors differently for now.

Bug: 421203760
Test: CommonTargetLanguageTest, AndroidX changes: aosp/3727966
Change-Id: I40650f845524cc7ab87c196557a66fa8310e6021