Update removed callable compatibility check for kotlin only items
As a followup to aosp/3721971, this handles the case when a kotlin only callable is removed, but there is still an overload which could be resolved from all existing usages of the old callable in kotlin code. An example of this is adding an optional parameter to a reified inline function. These functions can't be used from Java, and are inlined in all compiled usages, so as long as there is still an overload which is source compatible with the old version, this is a compatible change.

Bug: 245139957
Test: TargetLanguageCompatibilityTest
Change-Id: I3bf23359250bb3b4052a315d090c9d8dd97c0fb7