Require JvmName for file facade classes
When a file has top-level members which can be used by Java clients, it should use JvmName to provide the Java clients with a nice name.

This check doesn't count suspend functions, which were mentioned in the original bug, as those are generally not meant to target Java even though they technically can be called from Java.

The check also skips checking multi-file facade classes. It is possible to use JvmMultifileClass without JvmName, but it wouldn't make much sense to do so, and accessing the annotations of a multi-file class to check if has JvmName annotations isn't straightforward from psi.

Bug: 165356974
Test: KotlinInteropChecksTest
Change-Id: Ia5de6253600e6d3c580c30b0a5d5b46871d3072a