Push DefaultItem.appendDocumentation(...) down into subclasses
Previously, `DefaultItem.appendDocumentation(...)` would check to see
whether `this` was a `ParameterItem` and if it was then it would
delegate to the containing method's `appendDocumentation(...)` method.
This change pushes the implementation down into `DefaultParameterItem`
and `DefaultSelectableItem` to avoid that check.

Part of the work to clean up the `Item.documentation` property uses.

Bug: 429965593
Test: ./gradlew
Change-Id: I2979bc2ad1f450565e425d23f34f70b97b330697