Add test for using full version codes in RequiresApi
Currently, `DocAnalyzer.handleRequiresApi(...)` assumes that the value
supplied to `RequiresApi` was an API version level, i.e. a major
version. However, that is not completely true. It can also be a full
version code which encodes both major and minor. This change adds a
test to show the current, incorrect behavior. A follow-up change will
fix that.

Bug: 424435764
Test: ./gradlew
Change-Id: I6f5c3eec3ca2e9c633380e22a357135ce60940df