Move up deprecated status change check
This change moves the deprecated status change check in
checkFlaggedApiOnPreviouslyReleasedApi() to the top of the method.

this prepares the method for future changes to the modifiers change check logic
that requires that the deprecated status is already checked to avoid cases where the deprecated change is picked up in the modifiers check.

Bug: 429966896
Test: run FlaggedApiLintTest.kt

Change-Id: I9b90dc175952e699bd940b99de09a67ab95d4015