Report incorrect non-block use of @hide
Previously, incorrect non-block use of `@hide` would just work and not
give developers any feedback that what they were doing has been
deprecated. This change reports it as an issue using the
`WARNING_ERROR_WHEN_NEW` severity which should cause the
"Build Warnings Aye Aye Analyzer" to prevent new cases being added.

This issue will be removed after all the existing non-block uses have
been removed and support for non-block uses have been removed from
Metalava. That is because the performance cost of supporting non-block
uses is due to having to search the whole comment for `@hide` which
would still have to be done to support the issue reporting.

However, it may not be removed immediately after support has been
removed. It depends on whether developers still need the guidance on
what to do.

This baselines the test for Turbine as Turbine does not yet support
tracking location of documentation comments.

Bug: 429965593
Test: ./gradlew
Change-Id: I25415566a642f9caaeadf3962b20377347cd902d