Make DefaultItem.codebase final
Previously, `DefaultItem.codebase` was not `final` because it was
overridden by `PsiItem` to change the type to `PsiBasedCodebase`. That
ended up with two backing fields, both called `codebase` in classes
that implemented `PsiItem`. However, the `DefaultItem.codebase` field
was inaccessible. It also meant that the `codebase` field could not be
used in the `DefaultItem` initializer because it would call the getter
on `PsiBasedCodebase` which would access its `codebase` field which had
not yet been initialized.

This change makes the `codebase` field final and renames
`PsiItem.codebase` to `psiCodebase` to make it clearer. That will allow
the `DefaultItem` initializer to access the `codebase` field which is
needed in a follow-up change.

Bug: 429965593
Test: ./gradlew
Change-Id: Idbdb42cce6c4bd07dee0911cd2fd686017090b4b