Optimize initializeFromPsiElement(PsiElement)
Previously, it would check the `PsiElement` to see if it was one of a
variety of types. If one matched then it would usually extract the
documentation text from that but for `UElement` it could drop out and
check `PsiDocCommentOwner`. While it is theoretically possible for a
class to implement both, in practice they are completely separate. This
change switches the series of separate `if` statements into a single
`when` statement that avoids that.

As part of that it also removes some unnecessary `null` checks, dedups
defaulting to an empty string and inlines `firstNotNullOfOrNull(...)`
to simplify the code.

This simplifies follow-up changes that add support for tracking the
`FileLocation` of the `ItemDocumentation`.

Bug: 429965593
Test: ./gradlew
Change-Id: Ic39052eff809a01060a4586578a6364de714e26f