Add ItemDocumentation.print(PrintWriter) to write documentation
Previously, `StubWriter.appendDocumentation(...)` would write the
`ItemDocumentation` to the `PrintWriter`. This change moves that into
`ItemDocumentation.print(PrintWriter)` to make it easier to refactor in
future.

The intent is that in future `ItemDocumentation` will have an internal
model of the Javadoc that will handle printing, including fully
resolving references where needed. That will be common across all
source model providers, replacing the Psi specific functionality.

Bug: 429965593
Test: ./gradlew
Change-Id: I0e053d8b01c89d8a404a155a297762cc5ca829a0