Use JavadocParser in DefaultDocDescription
Previously, `JavadocParser` was used to parse the whole Javadoc comment
but part of that work (that splits it into a main description and block
tags each with their oew description) has been moved into
`DocCommentParser` for performance reasons. This change switches
`JavadocParser` to just parse the description part, i.e. a mixture of
text (which may include HTML tags) and inline doc tags.

It adds a `content` property to `DefaultDocDescription` which is lazily
initialized when requested and which uses `JavadocParser` to parse the
substring for the description. It then makes use of that in
`JavadocParserTest` to ensure that it works correctly.

Bug: 429965593
Test: ./gradlew
Change-Id: I38c7f62ed7c6fe28e31276b4ac9606827848c735