[Refactor readability] rename createReportFile to createApiReportFile
This CL:
* rename method & parameter
* rename apiWriter variable to use Visitor naming

Context:
The method naming is inconsistent. It implies that it creates generic
report files that are described the the description parameter. But at
the same time takes in an "apiFile" as input that usually refers to a
specific use case (signature files). The use of apiWriter and codebase
reinforces this link to the api.

I feel merging the naming convention will create a better compromise.

Also the apiWriter is not a writer it's a Visitor, does calling it a
writer is misleading.

Bug: 442395516
Test: FlaggedApiLintTest.kt
Change-Id: Ib034597a5f8a85d05698ca3bcaca6d0870d83f86