Allow removal of default parameter value when a compatible overload exists
If a parameter becomes not optional, that is generally a source breaking change, but it is possible that all calls to the old function might still work by resolving to a different overload of the same function. This reuses functionality from the check added in aosp/3724507 which found a compatible overload that could replace all calls to a removed function.

Bug: 362963888
Test: CompatibilityCheckTest
Change-Id: Ibee45945f9cb39db1be52e0e26906a956e9048b7