Stop changing API visibility based on VisibleForTesting
This annotation was previously special-cased to change API visibility, but that does not align with the policy of how the annotation should be used.

The previous behavior can be mostly replicated using `--hide-annotation` and `--show-annotation`, and the existing tests were updated to demonstrate that. The difference is that VisibleForTesting.PROTECTED will no longer make an element have protected visibility.

Bug: 443016431
Test: VisibleForTestingTest
Change-Id: I7d1e6efb8c6d2f117d569cde8cae3800c0c8fed5