Separate kotlin and bytecode methods when return types differ
Usually when a Kotlin function uses a value clsas type, the bytecode signature is mangled. However, there is at least one case when it isn't, and currently in that case the Kotlin and bytecode signatures are merged into one entry, even though they should have different return types.

This is also needed to prevent some Kotlin @JvmName functions from being merged with bytecode functions which are not the same definition in the followup change.

Bug: 445490139
Test: CommonTargetLanguageTest

Change-Id: I81d7738b0d2a441ed4fc3886690194c47f58d59e