[Refactor prep] Change FlaggedApiLint into a DelegatedVisitor
Before:
FlaggedApiLint -> ApiVisitor
After:
FlaggedApiLint -> DelegatedVisitor

This CL:
* Turn FlaggedApiLint into a DelegatedVisitor
* Update the Flagged api lint callsite to now use a FilteringApiVisitor
  that mimics ApiLint's ApiVisitor

Context:
FlaggedApiLint is moving away from the ApiVisitor pattern that accepts a
codebase and applies api filtering logic determining which items to
visit. FlaggedApi now only focuses on what to do with a Item rather which
item to check.

This makes it easier to use it with the signature file writers code
fragment filtering logic in later Cls

Notes:
* FlaggedAPILint still has a few dependencies on ApiVisitor. I opted to
  expose these dependencies so it can still have access to them. The
idea is to eventually remove ApiVisitor all together and have APILint
also be a delegateVisitor. At that point, this shared logic can be moved
to a utility class. But for now I feel it is best to leave that for a
later refactor

Bug: 442395516
Test: FlaggedApiLintTest.kt
Change-Id: I6439bcd9e1d48adeff074356d7bc8945a84ac3ac